backend/my_portfolio_chat_entry_point/requirements.txt
flask
openai
google-cloud-firestore
gunicorn
